+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [next]
 
+## 0.2.0
+ - added flavors: production, development, devp
+ - improved logger, url provider, repository
+
 ## 0.1.1
  - fixed http token updating
  - updated flutter to v1.12.13+hotfix.6 and Podfile